atividade 15 | Versão 03

+++ src/main/java/br/com/ifba/curso/Service/CursoService.java
@@ -3,43 +3,75 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
 package br.com.ifba.curso.Service;
-import br.com.ifba.curso.entity.Curso;
+import br.com.ifba.curso.exceptions.Exceptions;
 import java.util.List;
 import br.com.ifba.curso.dao.dao;
+import br.com.ifba.curso.entity.Curso;
 /**
  *
  * @author Casa
  */
 public class CursoService implements CursoServiceI {
       private final dao cursoDao = new dao();
+      Exceptions excecoes = new Exceptions();// injeção da clsse criada para uso de exceções
 
     @Override
-    public Curso saveCurso(Curso curso) {
+    public Curso saveCurso(Curso curso)throws RuntimeException {
+        
+        if(excecoes.capsuleisNull(curso)){
+        
+            throw new RuntimeException ("Dados do curso não preenchidos");  
+        
+        }/* else if(excecoes.idisEqual(curso.getId())){
+        
+        throw new RuntimeException ("ID já existente em outro objeto"); 
+            
+        }  */else{
+        
         return cursoDao.save(curso);
+        }
     }
 
     @Override
-    public Curso updateCurso(Curso curso) {
+    public Curso updateCurso(Curso curso)throws RuntimeException {
+         if(excecoes.capsuleisNull(curso)){
+        
+            throw new RuntimeException ("Dados do curso não preenchidos");  
+        
+        }  else{
+        
         return cursoDao.update(curso);
+        }
     }
 
     @Override
-    public void deleteCurso(Curso curso) {
-        cursoDao.delete(curso);
+    public void deleteCurso(Curso curso)throws RuntimeException {
+        
+        if(excecoes.capsuleisNull(curso)){
+        
+            throw new RuntimeException ("Dados do curso não preenchidos");  
+        
+        } else{
+        
+            cursoDao.delete(curso);
+        
+        }
+        
     }
 
     @Override
-    public List<Curso> getAllCursos() {
+    public List<Curso> getAllCursos()throws RuntimeException {
         return cursoDao.findAll();
     }
 
     @Override
-    public Curso getCursoById(Long id) {
+    public Curso getCursoById(Long id)throws RuntimeException{
+  
         return cursoDao.findById(id);
     }
 
     @Override
-    public List<Curso> getCursoByName(String name) {
+    public List<Curso> getCursoByName(String name)throws RuntimeException{
         return cursoDao.findByName(name);
     }
 }
